Make lod calc faster, expand home, change module layout

+++ include/rust_octree_bind.h
@@ -8,4 +8,4 @@
 using namespace open3d::geometry;
 
 std::shared_ptr<Octree> load_octree_from_file(rust::String file_name);
-rust::Vec<double> get_intersecting_node(std::shared_ptr<Octree> octree, Viewbox viewbox);
+rust::Vec<int64_t> get_intersecting_node(std::shared_ptr<Octree> octree, Viewbox viewbox);
